Move to max_listeners_spec
