--- packages/server/test/support/fixtures/projects/plugin-max-listeners/cypress/plugins/index.js
@@ -1,6 +1,0 @@
-const webpackPreprocessor = require('../../../../../../npm/webpack-preprocessor')
-
-module.exports = (on) => {
-  console.log('plugin ran')
-  on('file:preprocessor', webpackPreprocessor())
-}
